Refactored Views folder to prepare for RNWeb
- Moved types to types.ts to split implementation and types
- Moved declaration of SkiaViewApi to types
- Moved implementation (import from global) to api file
