+++ package/src/views/SkiaView.tsx
@@ -1,63 +1,17 @@
 import React from "react";
-import type { ViewProps } from "react-native";
 import { requireNativeComponent } from "react-native";
 
-import type { SkImage, SkRect } from "../skia";
+import type { SkRect } from "../skia";
 import type { SkiaReadonlyValue } from "../values";
 
-import type {
-  DrawMode,
-  RNSkiaDrawCallback,
-  NativeSkiaViewProps,
-} from "./types";
+import { SkiaViewApi } from "./api";
+import type { DrawMode, NativeSkiaViewProps, SkiaViewProps } from "./types";
 
 let SkiaViewNativeId = 1000;
 
 const NativeSkiaView = requireNativeComponent<NativeSkiaViewProps>(
   "ReactNativeSkiaView"
 );
-
-declare global {
-  var SkiaViewApi: {
-    invalidateSkiaView: (nativeId: number) => void;
-    makeImageSnapshot: (nativeId: number, rect?: SkRect) => SkImage;
-    setDrawCallback: (
-      nativeId: number,
-      callback: RNSkiaDrawCallback | undefined
-    ) => void;
-    setDrawMode: (nativeId: number, mode: DrawMode) => void;
-    registerValuesInView: (
-      nativeId: number,
-      values: SkiaReadonlyValue<unknown>[]
-    ) => () => void;
-  };
-}
-
-const { SkiaViewApi } = global;
-
-export interface SkiaViewProps extends ViewProps {
-  /**
-   * Sets the drawing mode for the skia view. There are two drawing
-   * modes, "continuous" and "default", where the continuous mode will
-   * continuously redraw the view, and the default mode will only
-   * redraw when any of the regular react properties are changed like
-   * sizes and margins.
-   */
-  mode?: DrawMode;
-  /**
-   * When set to true the view will display information about the
-   * average time it takes to render.
-   */
-  debug?: boolean;
-  /**
-   * Draw callback. Will be called whenever the view is invalidated and
-   * needs to redraw. This is either caused by a change in a react
-   * property, a touch event, or a call to redraw. If the view is in
-   * continuous mode the callback will be called 60 frames per second
-   * by the native view.
-   */
-  onDraw?: RNSkiaDrawCallback;
-}
 
 export class SkiaView extends React.Component<SkiaViewProps> {
   constructor(props: SkiaViewProps) {
