+++ package/src/views/api.ts
@@ -1,0 +1,7 @@
+import type { ISkiaViewApi } from "./types";
+
+declare global {
+  var SkiaViewApi: ISkiaViewApi;
+}
+
+export const { SkiaViewApi } = global;
