+++ package/src/views/types.ts
@@ -1,6 +1,8 @@
 import type { ViewProps } from "react-native";
 
+import type { SkImage, SkRect } from "../skia";
 import type { SkCanvas } from "../skia/Canvas";
+import type { SkiaReadonlyValue } from "../values";
 
 export type DrawMode = "continuous" | "default";
 
@@ -53,4 +55,42 @@ export type RNSkiaDrawCallback = (canvas: SkCanvas, info: DrawingInfo) => void;
 export interface ValueListener {
   addListener: (callback: () => void) => number;
   removeListener: (id: number) => void;
+}
+
+export interface ISkiaViewApi {
+  invalidateSkiaView: (nativeId: number) => void;
+  makeImageSnapshot: (nativeId: number, rect?: SkRect) => SkImage;
+  setDrawCallback: (
+    nativeId: number,
+    callback: RNSkiaDrawCallback | undefined
+  ) => void;
+  setDrawMode: (nativeId: number, mode: DrawMode) => void;
+  registerValuesInView: (
+    nativeId: number,
+    values: SkiaReadonlyValue<unknown>[]
+  ) => () => void;
+}
+
+export interface SkiaViewProps extends ViewProps {
+  /**
+   * Sets the drawing mode for the skia view. There are two drawing
+   * modes, "continuous" and "default", where the continuous mode will
+   * continuously redraw the view, and the default mode will only
+   * redraw when any of the regular react properties are changed like
+   * sizes and margins.
+   */
+  mode?: DrawMode;
+  /**
+   * When set to true the view will display information about the
+   * average time it takes to render.
+   */
+  debug?: boolean;
+  /**
+   * Draw callback. Will be called whenever the view is invalidated and
+   * needs to redraw. This is either caused by a change in a react
+   * property, a touch event, or a call to redraw. If the view is in
+   * continuous mode the callback will be called 60 frames per second
+   * by the native view.
+   */
+  onDraw?: RNSkiaDrawCallback;
 }
